Convert RollingAverage to int64
The commited version of the RollingAverage class uses doubles as the interior
type. This was a compromise since I was intending on using it for both floating
point and integer data. I have decided to not use it for the floating point
data, so I am converting it to the correct integer type. If there is a need in
the future for it to support multiple data types I would consider making it a
template class, but I think that is too heavy weight of a solution at the
moment.

BUG=chromium-os:23647
TEST=Confirmed that the code still builds and the unit tests pass.

Reviewed-on: https://gerrit.chromium.org/gerrit/17524
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>

+++ power_manager/rolling_average.h
@@ -20,12 +20,12 @@ class RollingAverage {
 
   virtual void Init(unsigned int max_window_size);
 
-  virtual double AddSample(double sample);
-  virtual double GetAverage();
+  virtual int64 AddSample(int64 sample);
+  virtual int64 GetAverage();
   virtual void Clear();
  protected:
   void DeleteSample();
-  void InsertSample(double sample);
+  void InsertSample(int64 sample);
   bool IsFull();
 
  private:
@@ -49,8 +49,8 @@ class RollingAverage {
   FRIEND_TEST(RollingAverageTest, IsFullUninitialized);
   FRIEND_TEST(RollingAverageTest, IsFullOverflow);
 
-  std::queue<double> sample_window_;
-  double running_total_;
+  std::queue<int64> sample_window_;
+  int64 running_total_;
   unsigned int max_window_size_;
 
   DISALLOW_COPY_AND_ASSIGN(RollingAverage);
